third_party/blink/web_tests/fast/frames/sandboxed-iframe-navigation-top-denied-expected.txt

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with origin 'navigate-top-to-fail.html'. The frame attempting navigation of the top-level window is sandboxed, but the flag of 'allow-top-navigation' or 'allow-top-navigation-by-user-activation' is not set.

CONSOLE ERROR: Uncaught SecurityError: Failed to set a named property 'href' on 'Location': The current window does not have permission to navigate the target frame to 'fail-and-notify-done.html'.
This test verifies that a sandboxed IFrame cannot navigate the top-level frame without allow-top-navigation. This test passes if the navigation does not occur.
